Add files via upload
The client would like to come see a dynamic dashboard with different KPI's at different 
levels (National, Region & Territory etc). Create the below charts as these would help us 
explore the past data in a better manner and give a good picture of the progress and 
failures. This in turn would catalyse the decision-making process, making it easier, simpler 
and accurate